long Id统一换为int Id(以前使用long是因为sqlite自增主键必须使用INTEGER(long),软件换为int后,sqlite可以继续使用long,兼容使用) 修正ToBool()
git-svn-id: https://tinn-pc:8443/svn/Common/Paway.Forms@731 c1914438-a9fc-4b41-8dc5-b03b8268c31d

+++ Paway.Forms/TreeGridView/TreeGridView.cs
@@ -83,16 +83,16 @@ namespace Paway.Forms
             foreach (var temp in tempList)
             {
                 var node = Nodes.Add(type.GetValue(temp));
-                AddNodes(type, list, node, (long)type.GetValue(temp, nameof(IParent.Id)));
+                AddNodes(type, list, node, (int)type.GetValue(temp, nameof(IParent.Id)));
             }
         }
-        private void AddNodes(Type type, IList list, TreeGridNode parent, long parentId)
+        private void AddNodes(Type type, IList list, TreeGridNode parent, int parentId)
         {
             var tempList = type.FindAll(list, nameof(IParent.ParentId), parentId);
             foreach (var temp in tempList)
             {
                 var node = parent.Nodes.Add(type.GetValue(temp));
-                AddNodes(type, list, node, (long)type.GetValue(temp, nameof(IParent.Id)));
+                AddNodes(type, list, node, (int)type.GetValue(temp, nameof(IParent.Id)));
             }
         }
         private void AutoColumns(Type type)
